[TPDE] Don't emit jump tables inside code
This avoids mixing data and instructions too hard and is permits using
different, more compact encoding schemes in the future.

On x86-64, jump tables are emitted at the end of the function. This is
different to what e.g. LLVM emits, but avoids emitting one relocation
per jump table entry.

On AArch64, jump tables are emitted into the rodata section.