[TPDE][NFC] Replace TPDE_ASSERTS macro with constexpr bool
Instead of messing with the NDEBUG macro in unexpected ways, add a CMake
option TPDE_ENABLE_ASSERTIONS that undefines NDEBUG again and thereby
permits assertions to be enabled in release builds.

Additionally, replace some usages of ifdef with constexpr if so that the
code is also parsed in release builds, preventing e.g. warnings.